doing 18900000 on the dash im not crashing

+++ Buttons/Default/exit.js
@@ -7,7 +7,7 @@ module.exports = {
    * @param {ButtonInteraction} interaction
    */
   async execute(interaction) {
-    if (interaction.message.deletable) {
+    if (interaction.message && interaction.message.deletable) {
       await interaction.message.delete().catch((err) => {
         return interaction.reply({
           content:
